requirements.txt
# ===== 采集 =====
uiautomator2==3.2.3
adbutils==2.8.0
Appium-Python-Client==4.0.0

# ===== 解析 =====
paddleocr==2.7.0.3
paddlepaddle==2.6.0
opencv-python==4.9.0.80
lxml==5.2.1
imagehash==4.3.1
Pillow==10.3.0

# ===== 存储 =====
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
redis==5.0.4
minio==7.2.5

# ===== 任务调度 =====
celery==5.3.6
APScheduler==3.10.4

# ===== API & Web =====
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3

# ===== 推送/渲染 =====
requests==2.31.0
pyppeteer==2.0.0

# ===== 工具 =====
pyyaml==6.0.1
pydantic==2.7.0
pandas==2.2.2
loguru==0.7.2
